docs(three-d-node-tree): rewrite stories in new format (#549)
